+++ validators/keyValidator.js
@@ -1,18 +1,62 @@
 const { checkSchema } = require("express-validator");
-const { getRoleKey } = require("../db/queries");
+const { keyExists } = require("../db/queries");
 
-const keyValidator = async (key) => {};
+const keyFormatValidator = (key) => {
+  console.log("keyFormatValidator running...");
+  const regex = new RegExp("([A-Z0-9]{4}(?:[- ]?[A-Z0-9]{4}){3})");
+  return regex.test(key);
+};
 
+const keyValidator = async (key, { req }) => {
+  console.log("keyValidator running...");
+  console.log("key:", key);
+  const { id: accountID } = req.user;
+  console.log("accountID:", accountID);
+  const result = await keyExists({ accountID, key });
+  console.log("result:", result);
+  if (!result) {
+    console.log("!result:", !result);
+    throw new Error();
+  }
+  return Promise.resolve();
+};
+
+const keySanitizer = (initialKey) => {
+  /* This sanitizer will only run if the previous validator passes
+   * This will sanitized key to XXXX-XXXX-XXXX-XXXX
+   * XXXXXXXXXXXXX, or XXXX XXXX XXXX XXXX => XXXX-XXXX-XXXX-XXXX
+   */
+  const regex = new RegExp("[0-9A-Z]{4}", "g");
+  const sanitizedKey = initialKey.match(regex).join("-");
+  return sanitizedKey;
+};
+
+/* const pattern = new RegExp('([A-Z0-9]{4}(?:[- ]?[A-Z0-9]{4}){3})');
+ * Test string is one of three formats: XXXX-XXXX-XXXX-XXXX, XXXXXXXXXXXXX, or XXXX XXXX XXXX XXXX
+ * If test returns true
+ *  Continue with validation
+ * Else
+ *  Break out of validation
+ * Sanitize string to the format XXXX-XXXX-XXXX-XXXX before validating against the database
+ */
 const keySchema = {
   key: {
     trim: true,
     isEmpty: {
       negated: true,
+      errorMessage: "Key cannot be empty.",
     },
-    errorMessage: "Key cannot be empty.",
-    custom: {
-      options: keyValidator,
+    isFormat: {
+      custom: keyFormatValidator,
+      bail: true,
     },
+    customSanitizer: {
+      options: keySanitizer,
+    },
+    isValid: {
+      custom: keyValidator,
+    },
+    errorMessage: "Invalid key.",
     escape: true,
   },
 };