+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 ---
+### 21 JAN 2025
+- The `keyExists` query function now accommodates whether or not the given key exists and if the account has activated the key's corresponding role.
+- Created `insertPost`, `getAccountPosts`, and `getAllPosts` query functions in the `queries.js` module.
+- Created `postForm.ejs` and `posts.ejs` partials.
+- 
+---
 ### 17 JAN 2025
 - Grouped `account_id` and `role_id` columns in `user_roles` table as a unique pair; an account can be in multiple roles, but cannot be in the same role multiple instances.
 - The values for the `activation_key` column are now a substring of `gen_random_uuid()::text` with the following format: `XXXX-XXXX-XXXX-XXXX`.